+++ utils/__init__.py
@@ -1,4 +1,6 @@
 from .commands import *
 from .decorators import *
+from .embed import *
 from .error import *
 from .logging import *
+from .view import *
